lesson 8: laravel ui, middleware, sessions

+++ app/Http/Controllers/Admin/NewsController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreNewsRequest;
+use App\Http\Requests\UpdateNewsRequest;
 use App\Models\Category;
 use App\Models\News;
 use Illuminate\Http\Request;
@@ -49,27 +51,22 @@ class NewsController extends Controller
 		]);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\RedirectResponse
+	/**
+	 * Store a newly created resource in storage.
+	 *
+	 * @param StoreNewsRequest $request
+	 * @return \Illuminate\Http\RedirectResponse
 	 */
-    public function store(Request $request)
+    public function store(StoreNewsRequest $request)
     {
-    	$request->validate([
-    		'title' => ['required', 'string'],
-		]);
-
-    	$data = $request->only(['category_id', 'title', 'description', 'author', 'status']);
-    	$news = News::create($data);
+    	$news = News::create($request->validated());
 
 		if($news) {
 			return redirect()->route('admin.news.index')
-				->with('success', 'Новость успешно добавлена');
+				->with('success', __('messages.admin.news.create.success'));
 		}
 
-		return back()->withInput()->with('error', 'Не удалось добавть новость');
+		return back()->withInput()->with('error', __('messages.admin.news.create.fail'));
     }
 
     /**
@@ -101,40 +98,40 @@ class NewsController extends Controller
 	/**
 	 * Update the specified resource in storage.
 	 *
-	 * @param \Illuminate\Http\Request $request
+	 * @param UpdateNewsRequest $request
 	 * @param News $news
 	 * @return \Illuminate\Http\RedirectResponse
 	 */
-    public function update(Request $request, News $news)
+    public function update(UpdateNewsRequest $request, News $news)
     {
-		$request->validate([
-			'title' => ['required', 'string'],
-		]);
-
-		$news = $news->fill(
-			$request->only(['category_id', 'title', 'description', 'author', 'status'])
-		)->save();
+		$news = $news->fill($request->validated())->save();
 
 		if($news) {
 			return redirect()->route('admin.news.index')
-				->with('success', 'Новость успешно сохранена');
+				->with('success', __('messages.admin.news.update.success'));
 		}
 
-		return back()->withInput()->with('error', 'Не удалось сохранить новость');
+		return back()->withInput()->with('error', __('messages.admin.news.update.fail'));
     }
 
 	/**
 	 * Remove the specified resource from storage.
 	 *
 	 * @param News $news
-	 * @return \Illuminate\Http\Response
+	 * @return \Illuminate\Http\JsonResponse
 	 */
-    public function destroy(News $news)
+    public function destroy(Request $request, News $news)
     {
-        try{
-        	$news->delete();
-		}catch (\Exception $e) {
+      if($request->ajax()) {
+		  try {
+			  $news->delete();
+			  return response()->json('ok');
 
-		}
+		  } catch (\Exception $e) {
+             \Log::error($e->getMessage());
+
+             return response()->json('error', 400);
+		  }
+	  }
     }
 }
